style: Rename components
NullDataBox -> BaseNullDataBox
PageHeader  -> BasePageHeader
Panel       -> BasePanel
