+++ src/components/BaseNullDataBox.vue
