+++ src/components/BasePanel.vue
